refactor: modularize UI into src package
